Update route nesting
