+++ web/app/components/product-link.hbs
@@ -1,5 +1,5 @@
 <LinkTo
-  @route="authenticated.product-areas"
+  @route="authenticated.product-areas.product-area"
   @model={{dasherize @product}}
   class="product-link"
   ...attributes
