--- web/app/controllers/authenticated/product-areas.ts
@@ -1,10 +1,0 @@
-import Controller from "@ember/controller";
-import AuthenticatedProductAreasRoute from "hermes/routes/authenticated/product-areas";
-import { ModelFrom } from "hermes/types/route-models";
-
-export default class AuthenticatedProductAreasController extends Controller {
-  queryParams = ["product"];
-  product = [];
-
-  declare model: ModelFrom<AuthenticatedProductAreasRoute>;
-}
