+++ web/app/styles/typography.scss
@@ -44,6 +44,8 @@
 .underlined-link:not([disabled]),
 .underline-on-hover:not([disabled]):hover,
 .underline-on-hover:not([disabled]):focus {
+  &:not(.product-abbreviation) {
+  }
   @apply text-color-foreground-faint underline decoration-color-palette-neutral-300 underline-offset-[3px];
 
   &:hover,
