+++ web/app/templates/authenticated/product-areas.hbs
@@ -1,7 +1,1 @@
-{{page-title this.model.productArea}}
-
-<ProductArea
-  @productArea={{this.model.productArea}}
-  @nbHits={{this.model.nbHits}}
-  @docs={{this.model.docs}}
-/>
+{{outlet}}
