+++ web/app/templates/authenticated/product-areas/product-area.hbs
@@ -1,0 +1,7 @@
+{{page-title this.model.productArea}}
+
+<ProductArea
+  @productArea={{this.model.productArea}}
+  @nbHits={{this.model.nbHits}}
+  @docs={{this.model.docs}}
+/>
